move tests to prevent parallel test execution
